backend/requirements.txt
fastapi
uvicorn
groq
python-dotenv
pydantic
duckduckgo_search
